requirements.txt
asgiref==3.7.2
cffi==1.16.0
channels==4.0.0
click==8.1.7
cryptography==42.0.5
dj-database-url==2.1.0
Django==5.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
h11==0.14.0
packaging==24.0
paho-mqtt==2.0.0
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
uvicorn==0.29.0
whitenoise==6.6.0
